+++ src/redux/store.js
@@ -3,11 +3,13 @@ import { configureStore } from '@reduxjs/toolkit'
 
 // In-house dependencies
 import alertsReducer from './slices/alertsSlice'
+import locationReducer from './slices/locationSlice'
 import notificationsReducer from './slices/notificationsSlice'
 
 const store = configureStore({
   reducer: {
     alerts: alertsReducer,
+    location: locationReducer,
     notifications: notificationsReducer,
   },
 })